requirements.txt
streamlit==0.74.0
opencv_python==4.5.1.48
numpy==1.19.2
matplotlib==3.3.3
Pillow==8.1.0
